+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes in keycloak-idp-mappers will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.1] - 2026-06-03
+
+### Changed
+- Enhance posix mapper description
+
 ## [2.1.0] - 2026-05-12
 
 ### Added